feat: dac prepare
